+++ refgenconf/const.py
@@ -4,7 +4,6 @@ Config file structure determination for the refgenie suite of packages
 These values are defined here in refgenconf and use some within this package,
 but they're also integral to both refgenie and to refgenieserver.
 """
-# config file structure related consts
 
 CFG_NAME = "genome configuration"
 CFG_ENV_VARS = ["REFGENIE"]
@@ -13,12 +12,14 @@ DEFAULT_SERVER = "http://refgenomes.databio.org"
 API_VERSION = "v2"
 DEFAULT_TAG = "default"
 
-
+# file or dir names
 TEMPLATE_RECIPE_JSON = "build_recipe_{}__{}.json"
 TEMPLATE_TARGET = "{}_{}__{}.flag"
 TEMPLATE_LOG = "build_log_{}__{}.md"
+ORI_LOG_NAME = "refgenie_log.md"
+BUILD_STATS_DIR = "_refgenie_build"
 
-FILENAME_TEMPLATES = ["TEMPLATE_RECIPE_JSON", "TEMPLATE_TARGET", "TEMPLATE_LOG"]
+FILE_DIR_NAMES = ["TEMPLATE_RECIPE_JSON", "TEMPLATE_TARGET", "TEMPLATE_LOG", "ORI_LOG_NAME", "BUILD_STATS_DIR"]
 
 # project-wide definition of the endpoint IDs. They are used to establish the way of communication between the server
 # and the client so that changes of endpoint function names OR endpoints themselves do not influence the connection
@@ -30,10 +31,22 @@ API_ID_ASSET_ATTRS = CUSTOM_PFX + "_asset_attrs"
 API_ID_DIGEST = CUSTOM_PFX + "_asset_digest"
 API_ID_RECIPE = CUSTOM_PFX + "_asset_recipe"
 API_ID_LOG = CUSTOM_PFX + "_asset_log"
+API_ID_ARCHIVE_DIGEST = CUSTOM_PFX + "_asset_archive_digest"
+API_ID_SPLASH = CUSTOM_PFX + "_asset_splash"
+
+# this dictionary groups the operationIds so that they can be accessed as modules
+# for systematic links generation in the splash pages
+OPERATION_IDS = {
+    "asset": {
+        API_ID_ARCHIVE: "archive", API_ID_ASSET_ATTRS: "attributes", API_ID_DIGEST: "asset digest",
+        API_ID_ARCHIVE_DIGEST: "archive digest", API_ID_RECIPE: "build recipe", API_ID_LOG: "build log"
+    }
+}
 
 API_IDS = [
     "API_ID_ASSETS", "API_ID_ARCHIVE", "API_ID_DEFAULT_TAG", "API_ID_ASSET_ATTRS", "API_ID_DIGEST",
-    "API_ID_RECIPE", "API_ID_LOG"]
+    "API_ID_RECIPE", "API_ID_LOG", "API_ID_SPLASH", "API_ID_ARCHIVE_DIGEST"
+]
 
 CFG_FOLDER_KEY = "genome_folder"
 CFG_SERVERS_KEY = "genome_servers"
@@ -80,10 +93,12 @@ REFGENIE_BY_CFG = {"0.3": "0.7.0", "0.2": "0.6.0"}
 ATTRS_COPY_PULL = [CFG_ASSET_DESC_KEY, CFG_SEEK_KEYS_KEY, CFG_ASSET_PARENTS_KEY, CFG_ASSET_PATH_KEY,
                    CFG_ASSET_CHECKSUM_KEY, CFG_TAG_DESC_KEY]
 REQ_TAG_ATTRS = [CFG_ASSET_PATH_KEY, CFG_SEEK_KEYS_KEY]
+CUSTOM_BAR_FMT = "{desc}{percentage:3.0f}%|{bar}| {n_fmt} [{elapsed}<{remaining}{rate_fmt}{postfix}]"
 
 __all__ = ["DEFAULT_SERVER", "CFG_ASSET_DEFAULT_TAG_KEY", "CFG_KEY_NAMES", "CFG_GENOME_DESC_KEY", "REQ_CFG_VERSION",
            "CFG_ASSETS_KEY", "CFG_GENOME_ATTRS_KEYS", "REFGENIE_BY_CFG", "DEFAULT_TAG", "ATTRS_COPY_PULL",
-           "REQ_TAG_ATTRS", "API_VERSION", "CONF_STRUCTURE"] + FILENAME_TEMPLATES + API_IDS + CFG_CONST + CFG_KEY_NAMES
+           "REQ_TAG_ATTRS", "CUSTOM_BAR_FMT", "API_VERSION", "CONF_STRUCTURE", "OPERATION_IDS", "CUSTOM_PFX"] + \
+          FILE_DIR_NAMES + CFG_CONST + CFG_KEY_NAMES + API_IDS
 
 CONF_STRUCTURE = """
 # example genome configuration structure